fix: lift CES client from per-session to server-level singleton
Managed CES accepts exactly one bootstrap connection then unlinks the
socket, so per-session CES initialization broke all conversations after
the first. Move CES lifecycle (process manager + client handshake) to
DaemonServer so one connection is shared across all sessions.
